@@ -4,23 +4,24 @@
   "private": true,
   "devDependencies": {
     "@types/jest": "^20.0.4",
-    "@types/react": "^15.0.35",
-    "@types/react-native": "^0.48.0",
-    "@types/react-navigation": "^1.0.15",
+    "@types/react": "^15.6.15",
+    "@types/react-native-vector-icons": "^4.6.0",
+    "@types/react-native": "^0.52.24",
+    "@types/react-navigation": "^1.5.5",
     "babel-jest": "^20.0.3",
-    "babel-preset-env": "^1.6.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2016": "^6.24.1",
     "babel-preset-node6": "^11.0.0",
     "jest": "~20.0.0",
-    "jest-fetch-mock": "^1.2.0",
+    "jest-fetch-mock": "^1.5.0",
     "mock-async-storage": "^1.0.3",
-    "react-native-debugger-open": "^0.3.11",
-    "react-native-typescript-transformer": "^1.1.3",
+    "react-native-debugger-open": "^0.3.17",
+    "react-native-typescript-transformer": "^1.2.4",
     "react-test-renderer": "16.0.0-alpha.12",
     "replace": "^0.3.0",
     "ts-jest": "^20.0.7",
-    "tslint": "^5.8.0",
-    "typescript": "^2.4.1"
+    "tslint": "^5.9.1",
+    "typescript": "^2.8.1"
   },
   "scripts": {
     "start": "react-native start",
@@ -31,7 +32,7 @@
     "lint": "tslint src/**/*.ts",
     "build": "npm run clean && npm run tsc --",
     "clean": "rm -rf build",
-    "postinstall": "npm run patch && rndebugger-open",
+    "postinstall": "npm run patch",
     "patch": "npm run patchNativeElements && npm run patchAndroidInstaller",
     "patchNativeElements": "./node_modules/.bin/replace '.*resizeMode.*' '' node_modules/react-native-elements/src/tile/Tile.js -s",
     "patchAndroidInstaller": "./node_modules/.bin/replace \\'install\\',\\ pathToApk \\'install\\',\\ \\'-r\\',\\ pathToApk node_modules/react-native/local-cli/runAndroid/runAndroid.js -s"
@@ -60,23 +61,22 @@
     "cacheDirectory": ".jest/cache"
   },
   "dependencies": {
-    "react": "16.0.0-alpha.12",
-    "react-native": "^0.48.1",
-    "react-native-easy-grid": "^0.1.13",
-    "react-native-easy-markdown": "^1.1.8",
-    "react-native-elements": "^0.15.0",
+    "react": "16.3.1",
+    "react-native": "^0.55.1",
+    "react-native-easy-grid": "^0.1.17",
+    "react-native-easy-markdown": "^1.1.9",
     "react-native-fetch-blob": "^0.10.8",
     "react-native-gravatar": "^1.0.2",
-    "react-native-image-picker": "^0.26.4",
+    "react-native-image-picker": "^0.26.7",
     "react-native-md5": "^1.0.0",
     "react-native-parse-html": "^2.2.6",
     "react-native-qrcode": "^0.2.4",
-    "react-native-settings-list": "^1.7.1",
+    "react-native-settings-list": "^1.8.0",
     "react-native-util": "^1.0.2",
-    "react-native-vector-icons": "^4.3.0",
+    "react-native-vector-icons": "^4.6.0",
     "react-native-xml2js": "^1.0.3",
-    "react-navigation": "^1.0.0-beta.11",
-    "tslib": "^1.7.1",
+    "react-navigation": "^1.5.11",
+    "tslib": "^1.9.0",
     "url": "^0.11.0",
     "util": "^0.10.3"
   }
